+++ lib/fortnox/structs/default_delivery_types.rb
@@ -3,13 +3,13 @@
 module Fortnox
   module Structs
     class DefaultDeliveryTypes < Fortnox::Struct
-      # Default delivery type for invoices. Can be PRINT EMAIL or PRINTSERVICE.
-      attribute? :invoice, Types::DefaultDeliveryTypeValues
+      # Default delivery type for invoices.
+      attribute? :invoice, Types::DefaultInvoiceDeliveryTypeValues
 
-      # Default delivery type for orders. Can be PRINT EMAIL or PRINTSERVICE.
+      # Default delivery type for orders.
       attribute? :order, Types::DefaultDeliveryTypeValues
 
-      # Default delivery type for offers. Can be PRINT EMAIL or PRINTSERVICE.
+      # Default delivery type for offers.
       attribute? :offer, Types::DefaultDeliveryTypeValues
     end
   end